refactored resolution validation into something cleaner
